优化 FileService 和 NoticeService 中的文件处理逻辑，增加空值检查，提升代码健壮性

+++ app-common/src/main/java/com/eryansky/modules/disk/service/FileService.java
@@ -153,7 +153,7 @@ public class FileService extends CrudService<FileDao, File> {
         String code = FileUploadUtils.encodingFilenamePrefix(fileName);
         String storePath = iFileManager.getStorePath(folder, sessionInfo.getUserId(), fileName);
 
-        String fileTemp = AppConstants.getDiskTempDir() + java.io.File.separator + code+ "_" + fileName;;
+        String fileTemp = AppConstants.getDiskTempDir() + java.io.File.separator + code + "_" + fileName;
         tempFile = new java.io.File(fileTemp);
         try (FileOutputStream fos = FileUtils.openOutputStream(tempFile)){
             IOUtils.copy(uploadFile.getInputStream(), fos);
@@ -204,6 +204,9 @@ public class FileService extends CrudService<FileDao, File> {
      */
     public void deleteByFileId(String fileId, boolean deleteDiskFile) {
         File file = dao.get(fileId);
+        if (file == null) {
+            return;
+        }
         try {
             //检查文件是否被引用
 //            List<File> files = this.findByCode(file.getCode(), fileId);
@@ -229,6 +232,9 @@ public class FileService extends CrudService<FileDao, File> {
      */
     public void deleteFileAndClearDisk(String fileId) {
         File file = dao.get(fileId);
+        if (file == null) {
+            return;
+        }
         try {
             iFileManager.deleteFile(file.getFilePath());
             logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
@@ -268,6 +274,9 @@ public class FileService extends CrudService<FileDao, File> {
      */
     public void clearFileAndClearDisk(String fileId) {
         File file = dao.get(fileId);
+        if (file == null) {
+            return;
+        }
         try {
             iFileManager.deleteFile(file.getFilePath());
             logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
@@ -306,6 +315,9 @@ public class FileService extends CrudService<FileDao, File> {
      */
     public void clearByFileId(String fileId) {
         File file = dao.get(fileId);
+        if (file == null) {
+            return;
+        }
         try {
             //检查文件是否被引用
             List<File> files = this.findByCode(file.getCode(), fileId);
@@ -325,33 +337,38 @@ public class FileService extends CrudService<FileDao, File> {
 
     /**
      * 根据文件夹ID级联删除文件（包含下级文件夹）
-     * @param fileId 文件ID
+     * @param folderId 文件夹ID
      * @return
      */
-    public void deleteCascadeByFolderId(String fileId) {
-        deleteCascadeByFolderId(fileId, false);
+    public void deleteCascadeByFolderId(String folderId) {
+        deleteCascadeByFolderId(folderId, false);
     }
 
     /**
      * 根据文件夹ID级联删除文件（包含下级文件夹）
-     * @param fileId 文件ID
-     * @param deleteDiskFile 删除磁盘文件
+     * @param folderId 文件夹ID
+     * @param deleteDiskFile 是否删除磁盘文件
      * @return
      */
-    public void deleteCascadeByFolderId(String fileId, boolean deleteDiskFile) {
-        File file = dao.get(fileId);
+    public void deleteCascadeByFolderId(String folderId, boolean deleteDiskFile) {
         try {
-            //检查文件是否被引用
-            List<File> files = this.findByCode(file.getCode(), fileId);
-            if (deleteDiskFile && Collections3.isEmpty(files)) {
-                iFileManager.deleteFile(file.getFilePath());
-                logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+            if (deleteDiskFile) {
+                List<File> files = findOwnerAndChildsFolderFiles(folderId);
+                for (File file : files) {
+                    try {
+                        iFileManager.deleteFile(file.getFilePath());
+                        logger.debug("删除文件：{}", new Object[]{file.getFilePath()});
+                    } catch (IOException e) {
+                        logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
+                    }
+                }
             }
+            File file = new File();
+            file.setFolderId(folderId);
+            file.preUpdate();
             dao.deleteCascadeByFolderId(file);
-        } catch (IOException e) {
-            logger.error("删除文件[{}]失败,{}", new Object[]{file.getFilePath(), e.getMessage()});
         } catch (Exception e) {
-            logger.error(e.getMessage());
+            logger.error(e.getMessage(), e);
         }
     }
